+++ doc/changelog.md
@@ -5,7 +5,9 @@
   - Add `docker.dockerFileDir` and `docker.dockerFile` to the properties configuration provider. (#438)
   - Fix splitting of bind volumes for Windows pathes (#443)
   - Add new build config option `user` for switching the user at the end of the Dockerfile. `docker.user` can be used
-    for the properties configuration provider (#441)    
+    for the properties configuration provider (#441)
+  - Include dot dirs when creating the build tar (#446)
+  
 * **0.15.1** (2016-05-03)
   - Fix push / pull progress bar ([#91](https://github.com/fabric8io/docker-maven-plugin/issues/91))
   - Allow empty environment variable ([#434](https://github.com/fabric8io/docker-maven-plugin/issues/434))